changed copy-to-clipboard-button and copyable-field to use copy-overlay this extracts duplicate code for better maintainability
